+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added ###
 
+- LinkAhead crawler for the metadata check.
+- Triggers the quality checker.
+- Docker build file and instructions.
+
 ### Changed ###
 
 ### Deprecated ###